// api/coreService/types/dto/JuzWithProgressBasicResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DtoJuzResponse } from "./JuzResponse.ts";
import type { DtoProgressHatamResponse } from "./ProgressHatamResponse.ts";

export type DtoJuzWithProgressBasicResponse = {
    /**
     * @type object | undefined
    */
    juz?: DtoJuzResponse;
    /**
     * @type object | undefined
    */
    progressHatam?: DtoProgressHatamResponse;
};