// api/coreService/types/dto/ProgressHatamResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DtoProgressHatamResponse = {
    /**
     * @type string | undefined
    */
    completedAt?: string;
    /**
     * @type string | undefined
    */
    createdAt?: string;
    /**
     * @type string | undefined
    */
    createdBy?: string;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type boolean | undefined
    */
    isCompleted?: boolean;
    /**
     * @type string | undefined
    */
    juzId?: string;
    /**
     * @type string | undefined
    */
    lastAyahId?: string;
    /**
     * @type number | undefined
    */
    progressPct?: number;
    /**
     * @type string | undefined
    */
    startAyahId?: string;
    /**
     * @type string | undefined
    */
    startedAt?: string;
    /**
     * @type string | undefined
    */
    updatedAt?: string;
    /**
     * @type string | undefined
    */
    updatedBy?: string;
    /**
     * @type string | undefined
    */
    userId?: string;
};